% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{clustering_determine_difference}
\alias{clustering_determine_difference}
\title{Distance Between Clusters}
\usage{
clustering_determine_difference(
  x_obs,
  y_obs,
  alldiff = NA,
  x_restructured = NA,
  clustering_method = NA
)
}
\arguments{
\item{x_obs}{A vector of indices representing the observations in the first
cluster}

\item{y_obs}{A vector of indices representing the observations in the second
cluster}

\item{alldiff}{A 2D array indicating the distance between each pair of
observations across all raters}

\item{x_restructured}{A 2D array containing the original data set, restructured to
facilitate more straightforward references in this function}

\item{clustering_method}{A string indicating the criterion that will be used
to evaluate the distance between clusters; permitted values are
\code{"average"}, \code{"minimum"}, \code{"maximum"}, \code{"median"},
\code{"centroid"}, and \code{"wald"}}
}
\value{
A numeric value indicating the distance between clusters
}
\description{
This helper function computes the distance between two clusters
  based on a specified metric.
}
