% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{compute_expected_pairwise_differences_lower_bound}
\alias{compute_expected_pairwise_differences_lower_bound}
\title{Compute Expected Pairwise Differences (Lower Bound)}
\usage{
compute_expected_pairwise_differences_lower_bound(
  x,
  shuffle_dimension,
  verbose
)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}

\item{shuffle_dimension}{A string indicating whether \code{"rows"} or
\code{"columns"} were shuffled in the original data set}

\item{verbose}{A boolean value indicating whether to provide progress updates}
}
\value{
A list of lists containing numeric vectors indicating the lower bound
  of the expected differences between pairs of raters
}
\description{
This helper function estimates the lower bound of the expected
  differences between pairs of raters.
}
