% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{compute_observed_differences}
\alias{compute_observed_differences}
\title{Compute Observed Differences}
\usage{
compute_observed_differences(x)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}
}
\value{
A two-element list indicating the observed differences between
  observations for all raters and for each pair of raters, respectively
}
\description{
This helper function computes the differences between all raters
  and between all pairs of raters in a given real or bootstrapped data set.
}
