% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{compute_single_expectation}
\alias{compute_single_expectation}
\title{Compute Single Expectation}
\usage{
compute_single_expectation(
  x,
  joint_alpha,
  pairwise_alpha,
  pairwise,
  samples = 1000,
  sampling_method = "nonparametric",
  verbose = FALSE
)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}

\item{joint_alpha}{A vector of concentration parameters representing all
raters}

\item{pairwise_alpha}{A list of lists of vectors of concentration parameters
representing all pairs of raters}

\item{pairwise}{A boolean value indicating whether pairwise differences
between individual raters should be computed; if \code{FALSE}, pairwise
differences will be ignored in order to reduce the time and memory
complexity of the computation}

\item{samples}{An integer value indicating how many samples to use to
estimate expected differences}

\item{sampling_method}{A string indicating whether bootstrapped sampling
should be performed using a \code{"parametric"} or \code{"nonparametric"}
approach}

\item{verbose}{A boolean value indicating whether to provide progress updates}
}
\value{
A two-element list indicating the expected differences between
  observations for all raters and for each pair of raters, respectively
}
\description{
This helper function computes the expected squared difference
  between deviates from the same underlying Dirichlet distribution when the
  categories in each distribution are not allowed to be reordered.
}
