% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{compute_standard_single_expectation}
\alias{compute_standard_single_expectation}
\title{Compute Standard Single Expectation}
\usage{
compute_standard_single_expectation(
  joint_alpha,
  pairwise_alpha,
  pairwise,
  verbose = FALSE
)
}
\arguments{
\item{joint_alpha}{A vector of concentration parameters representing all
cross-validation iterations}

\item{pairwise_alpha}{A list of lists of vectors of concentration parameters
representing all pairs of cross-validation iterations}

\item{pairwise}{A boolean value indicating whether pairwise reliability
between individual cross-validation iterations should be computed; if
\code{FALSE}, pairwise reliability will be ignored in order to reduce the
time and memory complexity of the computation}

\item{verbose}{A boolean value indicating whether progress updates should be
provided}
}
\value{
A two-element list indicating the expected difference between
  reordered Dirichlet deviates based on the concentration parameters given in
  \code{joint_alpha} and \code{pairwise_alpha}, respectively
}
\description{
This helper function computes the expected squared difference
  between deviates from the same underlying Dirichlet distribution when the
  categories in each distribution are not allowed to be reordered.
}
