% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{dirichlet.mle}
\alias{dirichlet.mle}
\title{Dirichlet.MLE}
\usage{
dirichlet.mle(
  x,
  weights = NULL,
  eps = 10^(-5),
  convcrit = 1e-05,
  maxit = 1000,
  oldfac = 0.3,
  verbose = FALSE
)
}
\arguments{
\item{x}{A list with each entry being a single cross-validation iteration;
each list entry should be either the output from a
\code{\link[topicmodels]{LDA}} function call or a 2D double vector with
distributions as the rows and each category within each distribution as the
columns (such that each row in the vector sums to 1)}

\item{weights}{A numeric vector used to calibrate the initial estimates of
concentration parameters}

\item{eps}{A numeric value used as a tolerance parameter to prevent
logarithms of zero}

\item{convcrit}{A numeric value indicating the threshold for convergence used
to estimate concentration parameters}

\item{maxit}{A numeric value indicating the maximum number of iterations used
to estimate concentration parameters}

\item{oldfac}{A numeric value between 0 and 1 used as the convergence
acceleration factor}

\item{verbose}{A boolean value indicating whether progress updates should be
provided}
}
\value{
A list of the estimated concentration parameters, the sum of those
  estimated concentration parameters, and the ratio between each estimated
  concentration parameter and the sum of those parameters
}
\description{
This helper function estimates the concentration parameters of
  the Dirichlet distribution underlying multiple deviates, with no
  restrictions placed on the values of those concentration parameters. This
  function is heavily based on \code{\link[sirt]{dirichlet.mle}}, with
  modifications to avoid potential singularities in the estimation procedure.
}
