% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{estimate_alpha_from_data}
\alias{estimate_alpha_from_data}
\title{Estimate Alpha from Data}
\usage{
estimate_alpha_from_data(
  x,
  pairwise,
  estimate_pairwise_alpha_from_joint,
  symmetric_alpha = FALSE,
  convcrit = 1e-05,
  maxit = 1000,
  verbose = FALSE
)
}
\arguments{
\item{x}{A list with each entry being a single cross-validation iteration;
each list entry should be either the output from a
\code{\link[topicmodels]{LDA}} function call or a 2D double vector with
distributions as the rows and each category within each distribution as the
columns (such that each row in the vector sums to 1)}

\item{pairwise}{A boolean value indicating whether pairwise reliability
between individual cross-validation iterations should be computed; if
\code{FALSE}, pairwise reliability will be ignored in order to reduce the
time and memory complexity of the computation}

\item{estimate_pairwise_alpha_from_joint}{A boolean value indicating, when
estimating reliability for pairs of cross-validation iterations, whether
each pair of iterations should use the concentration parameters estimated
across all iterations (\code{TRUE}) or whether separate sets of
concentration parameters should be estimated for each pair of
cross-validation iterations (\code{FALSE}); it is strongly suggested that
the default value of \code{TRUE} be used for this parameter}

\item{symmetric_alpha}{A boolean value indicating whether all concentration
parameters in each cross-validation iteration should be assumed to be
equal, which is common in LDA}

\item{convcrit}{A numeric value indicating the threshold for convergence used
to estimate concentration parameters}

\item{maxit}{A numeric value indicating the maximum number of iterations used
to estimate concentration parameters}

\item{verbose}{A boolean value indicating whether progress updates should be
provided}
}
\value{
A two-element list indicating the estimated concentration parameters
  for all concentration parameters and for each pair of cross-validation
  iterations, respectively
}
\description{
This helper function estimates the expected concentration
  parameters of the Dirichlet distribution underlying multiple deviates.
}
