% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{estimate_shuffled_single_expectation}
\alias{estimate_shuffled_single_expectation}
\title{Estimate Shuffled Single Expectation}
\usage{
estimate_shuffled_single_expectation(
  joint_alpha,
  pairwise_alpha,
  pairwise,
  samples = 1000,
  lower_bound = TRUE,
  method = "rotational",
  num_iter = (length(pairwise_alpha) + 1),
  zero2 = (10^(-255)),
  verbose = FALSE
)
}
\arguments{
\item{joint_alpha}{A vector of concentration parameters representing all
cross-validation iterations}

\item{pairwise_alpha}{A list of lists of vectors of concentration parameters
representing all pairs of cross-validation iterations}

\item{pairwise}{A boolean value indicating whether pairwise reliability
between individual cross-validation iterations should be computed; if
\code{FALSE}, pairwise reliability will be ignored in order to reduce the
time and memory complexity of the computation}

\item{samples}{An integer value indicating how many samples to use to
estimate expected differences when \code{shuffle=TRUE}}

\item{lower_bound}{A boolean value indicating whether the lower bound of the
expected differences (based on i.i.d. beta distributions) should be used
rather than shuffling Dirichlet distributions to estimate the exact
difference in every sample when \code{shuffle=TRUE}; for Dirichlet
distributions with many categories and function calls with a large value of
samples, setting this to \code{TRUE} is strongly recommended, as reordering
numerous samples with a large number of categories may require an
unreasonable amount of time}

\item{method}{A string indicating what method (\code{"rotational"} or
\code{"forward"}) will be used to reorder topics (see Britt, under review);
the \code{"forward"} method will be implemented at a later date}

\item{num_iter}{An integer value indicating the number of cross-validation
iterations}

\item{zero2}{A numeric value; in order to prevent errors when generating
Dirichlet deviates, any concentration parameters that are less than
\code{zero2} are changed to be \code{zero2}}

\item{verbose}{A boolean value indicating whether progress updates should be
provided}
}
\value{
A two-element list indicating the expected difference between
  reordered Dirichlet deviates based on the concentration parameters given in
  \code{joint_alpha} and \code{pairwise_alpha}, respectively
}
\description{
This helper function estimates the expected squared difference
  between deviates from the same underlying Dirichlet distribution when the
  categories in each distribution are allowed to be reordered.
}
