% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{estimate_single_expectation}
\alias{estimate_single_expectation}
\title{Estimate Single Expectation}
\usage{
estimate_single_expectation(
  x,
  joint_alpha,
  pairwise_alpha,
  pairwise,
  shuffle = TRUE,
  shuffle_method = "rotational",
  shuffle_dimension = "rows",
  clustering_method = "average",
  slow_clustering = FALSE,
  samples = 1000,
  sampling_method = "nonparametric",
  lower_bound = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}

\item{joint_alpha}{A vector of concentration parameters representing all
raters}

\item{pairwise_alpha}{A list of lists of vectors of concentration parameters
representing all pairs of raters}

\item{pairwise}{A boolean value indicating whether pairwise differences
between individual raters should be computed; if \code{FALSE}, pairwise
differences will be ignored in order to reduce the time and memory
complexity of the computation}

\item{shuffle}{A boolean value indicating whether or not the topics may have
been shuffled into different sequences by each rater}

\item{shuffle_method}{A string indicating what method (\code{"rotational"},
\code{"agglomerative"}, or \code{"divisive"}) will be used to reorder
topics}

\item{shuffle_dimension}{A string indicating whether \code{"rows"} or
\code{"columns"} should be reordered; this argument only has an effect if
\code{shuffle==TRUE} and \code{type==NA}}

\item{clustering_method}{A string indicating the criterion that will be used
to merge or divide clusters if \code{shuffle==TRUE} and either
\code{shuffle_method=="agglomerative"} or
\code{shuffle_method=="divisive"}; permitted values are \code{"average"},
\code{"minimum"}, \code{"maximum"}, \code{"median"}, \code{"centroid"}, and
\code{"wald"}}

\item{slow_clustering}{A boolean value indicating whether the distances
between clusters should be recalculated after each individual observation
is added to the new cluster (\code{TRUE}), which can improve the
cohesiveness of the resulting cluster, or whether all observations should
be added to the new cluster without recomputing the distances between
clusters after every addition (\code{FALSE}); this argument only has an
effect if \code{shuffle==TRUE} and \code{shuffle_method=="divisive"}}

\item{samples}{An integer value indicating how many samples to use to
estimate expected differences}

\item{sampling_method}{A string indicating whether bootstrapped sampling
should be performed using a \code{"parametric"} or \code{"nonparametric"}
approach}

\item{lower_bound}{A boolean value indicating whether the lower bound of the
expected differences (based on i.i.d. beta distributions) should be used
rather than reordering Dirichlet distributions to estimate the exact
difference in every sample when \code{shuffle==TRUE}}

\item{verbose}{A boolean value indicating whether to provide progress updates}
}
\value{
A two-element list indicating the expected differences between
  observations for all raters and for each pair of raters, respectively
}
\description{
This helper function uses bootstrapping to estimate the expected
  squared difference between observations from the same underlying
  distribution.
}
