% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{euclidkrip}
\alias{euclidkrip}
\title{Euclidean Krippendorff's Alpha}
\usage{
euclidkrip(
  x,
  pairwise = TRUE,
  force_bootstrapping = FALSE,
  shuffle = FALSE,
  shuffle_method = "rotational",
  shuffle_dimension = "rows",
  clustering_method = "average",
  slow_clustering = FALSE,
  samples = 1000,
  lower_bound = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}

\item{pairwise}{A boolean value indicating whether pairwise reliability
between individual raters should be computed; if \code{FALSE}, pairwise
reliability will be ignored in order to reduce the time and memory
complexity of the computation}

\item{force_bootstrapping}{A boolean value indicating whether bootstrapping
should be used to estimate expected differences rather than exactly
computing them even if \code{shuffle==FALSE}}

\item{shuffle}{A boolean value indicating whether or not the topics may have
been shuffled into different sequences by each rater; if this is
\code{TRUE} and \code{shuffle_dimension="rows"}, then the observations
(rows) corresponding to each rater will be reordered to achieve a (local)
optimal fit, whereas if this is \code{TRUE} and
\code{shuffle_dimension="columns"}, the categories (columns) will be
reordered instead}

\item{shuffle_method}{A string indicating what method (\code{"rotational"},
\code{"agglomerative"}, or \code{"divisive"}) will be used to reorder
topics}

\item{shuffle_dimension}{A string indicating whether \code{"rows"} or
\code{"columns"} should be reordered; this argument only has an effect if
\code{shuffle==TRUE} and \code{type==NA}}

\item{clustering_method}{A string indicating the criterion that will be used
to merge or divide clusters if \code{shuffle==TRUE} and either
\code{shuffle_method=="agglomerative"} or
\code{shuffle_method=="divisive"}; permitted values are \code{"average"},
\code{"minimum"}, \code{"maximum"}, \code{"median"}, \code{"centroid"}, and
\code{"wald"}}

\item{slow_clustering}{A boolean value indicating whether the distances
between clusters should be recalculated after each individual observation
is added to the new cluster (\code{TRUE}), which can improve the
cohesiveness of the resulting cluster, or whether all observations should
be added to the new cluster without recomputing the distances between
clusters after every addition (\code{FALSE}); this argument only has an
effect if \code{shuffle==TRUE} and \code{shuffle_method=="divisive"}}

\item{samples}{An integer value indicating how many samples to use to
estimate expected differences when \code{shuffle==TRUE}}

\item{lower_bound}{A boolean value indicating whether the lower bound of the
expected differences (based on i.i.d. beta distributions) should be used
rather than shuffling Dirichlet distributions to estimate the exact
difference in every sample when \code{shuffle==TRUE}}

\item{verbose}{A boolean value indicating whether \link{euclidkrip} and its
helper functions should provide progress updates beyond the estimation of
concentration parameters}
}
\value{
A list of class \code{euclidkrip_rel} containing six elements:
  Euclidean Krippendorff's alpha for each observation and all raters (as a
  numeric vector indicating the reliability for each observation), Euclidean
  Krippendorff's alpha for each observation and pairs of raters (as a list
  containing lists containing numeric vectors indicating the reliability of
  each observation for each pair of raters, e.g., the second element of the
  first list contains the reliability for raters 1 and 2), Euclidean
  Krippendorff's alpha for all observations and all raters (as a numeric
  value), Euclidean Krippendorff's alpha for all observations and pairs of
  raters (as a list containing lists containing numeric values, e.g., the
  second element of the first list contains the reliability for raters 1 and
  2), the matrix of reordered topics for each rater (if
  \code{shuffle==TRUE}), and any warnings raised
}
\description{
This function computes Euclidean Krippendorff's alpha for data
sets composed of compositional or Euclidean data, as described by Britt
(under review).
}
\details{
One of the most important uses for Euclidean Krippendorff's alpha is
to assess the reliability of the allocations of words to topics or of topics
to documents via topic modeling techniques that do not necessarily adhere to
a Dirichlet distribution. This includes, for instance, Dirichlet multinomial
mixture models, which are commonly used for short text topic modeling.

Crucially, different topic modeling cross-validation iterations, which
effectively represent distinct raters of the same data set, may result in the
same topics appearing in a different sequence. As such, those topics, which
may represent either the rows or the columns of the data set, often need to
be reordered in order to yield an optimal fit.

In practice, it is generally infeasible to assess every possible combination
of topics across all raters. As such, \link{euclidkrip} provides three
methods of converging toward an optimal topic order for each rater. The
default, \code{shuffle_method="rotational"}, takes the provided sequence and
swaps pairs of categories until no further swaps would further improve the
fit. \code{shuffle_method="agglomerative"} and
\code{shuffle_method="divisive"} instead perform hierarchical cluster
analyses of all topics, with the added restriction that two topics
constructed by the same rater may not appear in the same cluster. Notably,
the \code{"agglomerative"} option can take excessive time in some cases,
while the \code{"divisive"} option forms clusters based on the macro-level
dynamics of the data and may therefore yield weakly matched sets of
individual topics compared to the other available options.

Regardless of whether the rows or columns of the data set must be reordered,
when computing Euclidean Krippendorff's alpha, it is common to compute two
separate coefficients in this family: Euclidean Krippendorff's alpha for all
raters and for pairs of raters. Unlike Britt's nu, both of these Euclidean
Krippendorff's alpha measures yield a single reliability score for all
observations, as the formula used by Euclidean Krippendorff's alpha to
compute the expected differences between raters does not produce separate
results for each individual observation. However, if the reliability of each
observation is desired, then nonparametric bootstrapping may be used to
estimate those coefficients. This is automatically done when
\code{shuffle==TRUE}, and it is also implemented when
\code{force_bootstrapping==TRUE}. In either of those cases, the
single-observation reliability values will be reported alongside the
multiple-observation reliability values that \link{euclidkrip} already
provides.

By default, \link{euclidkrip} provides reliability coefficients for all
raters and for pairs of raters. To reduce the time and memory required, the
pairwise versions may optionally be omitted from the computation by
specifying \code{pairwise=FALSE}.

Additional usage notes:

1. \code{x} must be a list of 2D double vectors, and each row of each list
element should be a single observation from a Dirichlet distribution (e.g.,
an LDA topic) while each column should represent a category within that
distribution (e.g., a word).

2. When \code{shuffle==TRUE}, each element in \code{x} comprises a large
number of topics, and/or \code{samples} is large, setting
\code{lower_bound=TRUE} may sometimes be useful in order to eliminate the
need to reorder a large number of topics in numerous bootstrapped samples.
You may also wish to set \code{verbose=TRUE} to receive periodic progress
updates and ensure that the computation is proceeding as expected.

3. When \code{shuffle==TRUE}, the topics in \code{x} are reordered. This is
described in the matrix of reordered topics, which is provided as one of the
elements of the list returned by \link{euclidkrip} and can be viewed using
\code{summary()} on that object. Each row of this matrix indicates the manner
in which the topics were reordered. For instance, if the first row is
c(3,1,2), that means that for the first rater, the third topic was moved to
the first position, the first topic was moved to the second position, and the
second topic was moved to the third position.
}
\section{References}{

  Britt, B. C. (under review). Interrater reliability for compositional,
  Euclidean, and Dirichlet Data.
}

\examples{
rater1 <- rbind(c(0.80,0.05,0.15), c(0.10,0.85,0.05),
                c(0.05,0.05,0.90), c(0.85,0.10,0.05))
rater2 <- rbind(c(0.10,0.85,0.05), c(0.10,0.10,0.80),
                c(0.85,0.05,0.10), c(0.15,0.80,0.05))
data <- list(rater1, rater2)
rel <- euclidkrip(data, shuffle=TRUE, shuffle_method="agglomerative",
                  shuffle_dimension="columns", clustering_method="average",
                  samples=500, verbose=TRUE)
summary(rel)
summary(rel, element="all")
}
