% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{generate_artificial_data_nonparametric}
\alias{generate_artificial_data_nonparametric}
\title{Generate Artificial Data (Nonparametric)}
\usage{
generate_artificial_data_nonparametric(x, shuffle, shuffle_dimension)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}

\item{shuffle}{A boolean value indicating whether or not the topics may have
been shuffled into different sequences by each rater}

\item{shuffle_dimension}{A string indicating whether \code{"rows"} or
\code{"columns"} were shuffled in the original data set; this argument only
has an effect if \code{shuffle==TRUE}}
}
\value{
A list of nonparametrically bootstrapped data such that each element
  is a 2D double vector with observations as the rows and each category
  within each observation as the columns
}
\description{
This helper function uses nonparametric bootstrapping to
  generate an artificial data set from a provided data set
}
