% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{generate_artificial_data_parametric}
\alias{generate_artificial_data_parametric}
\title{Generate Artificial Data (Parametric)}
\usage{
generate_artificial_data_parametric(x, pairwise_alpha)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}

\item{pairwise_alpha}{A list of lists of vectors of concentration parameters
representing all pairs of raters}
}
\value{
A list of parametrically bootstrapped data such that each element is
  a 2D double vector with observations as the rows and each category within
  each observation as the columns
}
\description{
This helper function uses parametric bootstrapping to generate
  an artificial data set from a provided data set
}
