% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{generate_dirichlet_deviate_diffs}
\alias{generate_dirichlet_deviate_diffs}
\title{Generate Squared Dirichlet Deviate Differences}
\usage{
generate_dirichlet_deviate_diffs(x, samples, zero2 = (10^(-255)))
}
\arguments{
\item{x}{The concentration parameters used to generate Dirichlet deviates}

\item{samples}{An integer value indicating how many samples to use to
compute the sum of squared differences}

\item{zero2}{A numeric value; in order to prevent errors when generating
Dirichlet deviates, any concentration parameters that are less than
\code{zero2} are changed to be \code{zero2}}

\item{method}{A string indicating what method (\code{"rotational"} or
\code{"forward"}) will be used to reorder topics (see Britt, under review);
the \code{"forward"} method will be implemented at a later date}
}
\value{
A numeric value representing the sum of squared differences between
  Dirichlet deviates for \code{samples} times the number of pairs of deviates
}
\description{
This helper function generates two deviates from the same
  underlying Dirichlet distribution and returns the sum of the squared
  differences between the categories of those deviates.
}
