% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{label_assignments}
\alias{label_assignments}
\title{Label Assignments Matrix}
\usage{
label_assignments(assignments, type = NA)
}
\arguments{
\item{assignments}{A matrix indicating the relationship between each
reordered topic and its original position in the data}

\item{type}{A string indicating the type of reliability being assessed, with
permitted values of \code{"wt"} (word-topic reliability), \code{"td"}
(topic-document reliability), and NA}
}
\value{
A matrix indicating the relationship between each reordered topic and
  its original position in the data, with the addition of row and column
  labels
}
\description{
This helper function adds labels to the assignments matrix based
  on whether the rows and columns represent documents and topics, topics and
  words, or something else, in order to facilitate more informative printing.
}
