% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{restricted_divisive_clustering}
\alias{restricted_divisive_clustering}
\title{Restricted Divisive Clustering}
\usage{
restricted_divisive_clustering(
  x,
  verbose = FALSE,
  clustering_method = "average",
  slow_clustering = FALSE
)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector, with the topics
to be reordered representing the rows of each vector}

\item{verbose}{A boolean value indicating whether to provide progress updates}

\item{clustering_method}{A string indicating the criterion that will be used
to merge or divide clusters; permitted values are \code{"average"},
\code{"minimum"}, \code{"maximum"}, \code{"median"}, \code{"centroid"}, and
\code{"wald"}}

\item{slow_clustering}{A boolean value indicating whether the distances
between clusters should be recalculated after each individual observation
is added to the new cluster (\code{TRUE}), which can improve the
cohesiveness of the resulting cluster, or whether all observations should
be added to the new cluster without recomputing the distances between
clusters after every addition (\code{FALSE})}
}
\value{
A two-element list containing the reordered data set and a matrix
  indicating the relationship between each reordered topic and its original
  position in the data, respectively
}
\description{
This helper function performs divisive hierarchical cluster
  analysis in order to minimize the sum of squared differences between
  raters.
}
\details{
This procedure conducts a cluster analysis in order to group topics
  from each rater into clusters indicating optimal matches between those
  topics. This effectively serves to reorder the topics constructed by each
  such that, to the extent possible, each individual topic is matched with
  its best-fitting counterparts constructed by each other rater before
  reliability is assessed.

  This cluster analysis uses an additional constraint: all clusters resulting
  from dividing an existing cluster must contain a number of observations
  that is a multiple of the number of topics constructed by each rater, and
  all raters must be equally represented in each of those resulting clusters.
  This ensures that at the end of the process, the number of clusters will be
  equal to the number of topics constructed by each rater, and each rater
  will have yielded one topic belonging to each cluster.

  However, this approach sometimes creates clusters based on macro-level
  dynamics rather than micro-level differences between topics, ultimately
  yielding weakly matched sets of categories compared with other approaches.
  In those cases, either \link{rotational_shuffling} or
  \link{restricted_agglomerative_clustering} may yield superior results.
}
