% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{rotational_shuffling}
\alias{rotational_shuffling}
\title{Rotational Shuffling}
\usage{
rotational_shuffling(x, verbose = FALSE, swap_tol = (10^(-12)))
}
\arguments{
\item{x}{A list such that each element is a 2D double vector, with the topics
to be reordered representing the rows of each vector}

\item{verbose}{A boolean value indicating whether to provide progress updates}

\item{swap_tol}{A numeric value representing the minimum threshold by which
the sum of squared differences must be improved in order to execute a
potential swap of two rows; this is necessary in order to prevent infinite
loops due to minuscule rounding errors}
}
\value{
A two-element list containing the reordered data set and a matrix
  indicating the relationship between each reordered topic and its original
  position in the data, respectively
}
\description{
This helper function reorders the rows of ratings from multiple
  raters in order to minimize the sum of squared differences between them.
}
