% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{sirt_digamma1}
\alias{sirt_digamma1}
\title{Sirt Digamma 1}
\usage{
sirt_digamma1(x, h = 0.001)
}
\arguments{
\item{x}{A vector of concentration parameters}

\item{h}{A numeric value used to define a pair of nearby observations}
}
\value{
A list of the estimated concentration parameters, the sum of those
  estimated concentration parameters, and the ratio between each estimated
  concentration parameter and the sum of those parameters
}
\description{
This helper function estimates the derivative of the digamma
  function. This function is directly drawn from
  \code{\link[sirt]{sirt_digamma1}}, as published at
  \url{https://github.com/cran/sirt/blob/master/R/sirt_digamma1.R}.
}
