% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{squarediff}
\alias{squarediff}
\title{Sum of Squared Differences}
\usage{
squarediff(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
A numeric value representing the sum of the squared differences
  between \code{x} and \code{y}
}
\description{
This helper function computes the sum of squared differences
  between two vectors.
}
