% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{summary.brittnu_rel}
\alias{summary.brittnu_rel}
\title{Summary (brittnu)}
\usage{
\method{summary}{brittnu_rel}(object, element = NA, ...)
}
\arguments{
\item{object}{An object of class \code{brittnu_rel} outputted from
\link{brittnu}}

\item{element}{The specific reliability coefficient to be printed; permitted
values are \code{"singleomnibus"}, \code{"singlepairwise"},
\code{"multipleomnibus"}, \code{"multiplepairwise"}, \code{"all"}, and
\code{NA}}

\item{...}{Other arguments inherited from the generic \code{summary} function}
}
\description{
This method prints a summary of the Britt's nu reliability
  values using an object of class \code{brittnu_rel} outputted from
  \link{brittnu}.
}
