% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{summary.euclidkrip_rel}
\alias{summary.euclidkrip_rel}
\title{Summary (euclidkrip)}
\usage{
\method{summary}{euclidkrip_rel}(object, element = NA, ...)
}
\arguments{
\item{object}{An object of class \code{euclidkrip_rel} outputted from
\link{euclidkrip}}

\item{element}{The specific reliability coefficient to be printed; permitted
values are \code{"singleomnibus"}, \code{"singlepairwise"},
\code{"multipleomnibus"}, \code{"multiplepairwise"}, \code{"all"}, and
\code{NA}}

\item{...}{Other arguments inherited from the generic \code{summary} function}
}
\description{
This method prints a summary of the Euclidean Krippendorff's
  alpha reliability values using an object of class \code{euclidkrip_rel}
  outputted from \link{euclidkrip}.
}
