% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{estimate_alpha_from_data}
\alias{estimate_alpha_from_data}
\title{Estimate Alpha from Data}
\usage{
estimate_alpha_from_data(
  x,
  pairwise,
  estimate_pairwise_alpha_from_joint,
  symmetric_alpha = FALSE,
  convcrit = 1e-05,
  maxit = 1000,
  progress = FALSE
)
}
\arguments{
\item{x}{A list such that each element is a 2D double vector with
observations as the rows and each category within each observation as the
columns}

\item{pairwise}{A boolean value indicating whether pairwise differences
between individual raters should be computed; if \code{FALSE}, pairwise
differences will be ignored in order to reduce the time and memory
complexity of the computation}

\item{estimate_pairwise_alpha_from_joint}{A boolean value indicating, when
estimating reliability for pairs of raters, whether each pair of raters
should use the concentration parameters estimated across all raters
(\code{TRUE}) or whether separate sets of concentration parameters should
be estimated for each pair of raters (\code{FALSE}); this argument only has
an effect when \code{alpha==NA}, and it is strongly suggested that the
default value of \code{TRUE} be used for this argument when possible}

\item{symmetric_alpha}{A boolean value indicating whether concentration
parameters should be assumed to be unchanged between observations, which is
common in many topic modeling procedures}

\item{convcrit}{A numeric value indicating the threshold for convergence used
to estimate concentration parameters}

\item{maxit}{A numeric value indicating the maximum number of iterations used
to estimate concentration parameters}

\item{progress}{A boolean value indicating whether to provide progress
updates}
}
\value{
A two-element list indicating the estimated concentration parameters
  for all raters and for each pair of raters, respectively
}
\description{
This helper function estimates the expected concentration
  parameters of the Dirichlet distribution underlying multiple observations.
}
