% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brittnu.R
\name{shuffle_distributions}
\alias{shuffle_distributions}
\title{Shuffle Distributions}
\usage{
shuffle_distributions(x, method = "rotational", verbose = FALSE)
}
\arguments{
\item{x}{A list with each entry being a single cross-validation iteration;
each list entry should be either the output from a
\code{\link[topicmodels]{LDA}} function call or a 2D double vector with
distributions as the rows and each category within each distribution as the
columns (such that each row in the vector sums to 1)}

\item{method}{A string indicating what method (\code{"rotational"} or
\code{"forward"}) will be used to reorder topics (see Britt, under review);
the \code{"forward"} method will be implemented at a later date}

\item{verbose}{A boolean value indicating whether progress updates should be
provided}
}
\value{
A two-element list containing the reordered data set and a matrix
  indicating the relationship between each reordered category and its orginal
  position in the data
}
\description{
This helper function reorders the categories in multiple
  Dirichlet deviates in order to minimize the sum of squared differences
  between them.
}
